Scrape: Remove lists.tlug.jp/ML.[0-9] on exit
As explained in the README and the comment, these are copies of
lists.tlug.jp/index.html that are not actually on the site but
appear to be produced by a bug.

We use a shell `trap` to remove them on exit so that the command
runs even when wget fails. wget's error code should still be
returned by the script.