Scrape: Set --backups=0 to try to fix creation of `ML.1` "bug"
In the default configuration, when a file is re-fetched that would
overwrite an existing file on disk, the existing file is renamed with
a `.1` extension, the previous `.1` to `.2`, and so on.

It appears that, if a fetch of `foo/` resulted in saving `foo/index.html`,
the backup files are named not `index.html.1` but `foo.1`.

Unfortunately, this does not seem to fix the problem. But it does seem
like a good setting to keep anyway. Perhaps combined with other options
it will help.