Scrape: Explain problem with -N; expand --mirror to its component options
The manpage describes --mirror as a short way of just specifying a set
of several other options: `-r -N -l inf --no-remove-listing`.
Expanding this makes it easier to see exactly what's being done and
follow the (tricky) interactions of the various options as described
throughout the manpage. We use the short forms because, though not as
easy to read here, that's the form used in the manpage when describing
interactions between options.

We leave out --no-remove-listing because that applies only to FTP
retrievals, which we don't do here.

We use a trick for inline comments in Bash: `$(: "comment")`.

Also, document the problem with -N: our timestamps are by default the
time of the working copy checkout, not when the files were retrieved,
which means that in a fresh checkout all files will be considered at
least as new as what's on the site at that time. We'll have to think
about ways of dealing with this efficiently.