Scrape: Download, but do not commit, /MLadm/
The content under /MLadm/ looks sensitive, since it lightly
password-protected. (The login name and password are in graphic
images on the top page.)

Though we now download /MLadm/, it's not commited and is in the
.gitignore to avoid accidental commits. Thus, developers will
have a local copy, but GitLab will not be serving these files
over its web interface, nor will Netlify or other hosting be
serving it when it serves this repo. The name and password are
also obfuscated in the Scrape script.

To do this we needed to ignore robots.txt, unfortunately.
See the code for details.